// Presentation/Client/src/app/user/user.ts
export class User {
  firstName: string | undefined;
  middleName: string | undefined;
  lastName: string | undefined;
  birthDate: Date | undefined;
  mobileNumber: string | undefined;
  email: string | undefined;
  cityId: number | undefined; 
  governorateId: number | undefined;
  hasAddress: boolean = false;
  buildingNumber:string | undefined ;
  flatNumber:string | undefined;
  street:string | undefined;
}